chore: rename sketch parser module
