+++ packages/sktechapp-parser/ng-package.json
