+++ packages/sktechapp-parser/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sketch-parser",
+  "name": "@xlayers/sktechapp-parser",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^7.1.0",
