+++ packages/sktechapp-parser/src/lib/sketch-style-parser.service.spec.ts
