+++ packages/sktechapp-parser/tslint.json
